Revert "fix bug - use bigint for telegram chat_id"
This reverts commit 53a7c3fcc53ff31089cffefb94579335fa6a874c.

+++ app.py
@@ -40,7 +40,7 @@ def init_db():
         db['repos'].create_column('url', db.types.text)
         db.create_table('chats')
         db['chats'].create_column('token', db.types.string(40))
-        db['chats'].create_column('chat_id', db.types.bigint)
+        db['chats'].create_column('chat_id', db.types.integer)
         db.commit()
 
 @app.teardown_appcontext
